Day 3 solution
The Day 3 solution took about 15 mins to conceive of, and 2-3 hours to
code up. Ha. Tested my increasing understanding of maps and streams
here. The solution is super inefficient as I keep loading the file, I
could factor this out but even with this in, the final calculation for
part 2 took <5s.

Nice thing here was that part 2 very easily extended part 1, given the
way I'd done it. Part 2 took about 5 mins, which is a total novel
experience for me.