Code the Day 4 solution
Finished this almost a year after starting but it was relatively easy to
pick up which is nice/weird. Learned a bit along the way about using
regexes for data validation.

There are almost certainly opportunities for improvements here but I'm
not in a position to work them out right now. Some things that come to
mind:

* The validation functions are all written separately. This makes a lot
of sense to me in terms of readable code but there's probably some
shortcut to make it syntactically sweet.
* I'm munging together quite a lot with maps and counts, could probably
use reduces more effectively here.
* I handle cases with true and false, feels a little weird maybe.