Day 6 solution
Spread over a few days so harder to judge how long it took but I think
it's fair to say that the idea for the solution took about 20 mins, and
coding it up into something that worked took around 4-5 hours.

The idea was sound: find out how many people are in the group and look
for answers that were answered that many times. But structures were
    trick for me.

Anyways still pleased I got there.